# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "L5 OpenApi",
        "description": "L5 Swagger OpenApi description",
        "contact": {
            "email": "muhammad.akbar5999@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/orders": {
            "post": {
                "tags": [
                    "Orders"
                ],
                "summary": "Place a new order",
                "description": "Place a new order, and move all items in the cart to the order",
                "operationId": "placeOrder",
                "responses": {
                    "201": {
                        "description": "Order placed successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "isSuccess": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Order placed successfully"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Order"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "isSuccess": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Internal server error"
                                        },
                                        "data": {
                                            "example": null
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Order": {
                "title": "Order",
                "description": "Order model",
                "properties": {
                    "id": {
                        "description": "ID of the order",
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "user_id": {
                        "description": "ID of the user who placed the order",
                        "type": "integer",
                        "example": 1
                    },
                    "status": {
                        "description": "Status of the order",
                        "type": "string",
                        "example": "pending"
                    },
                    "total_price": {
                        "description": "Total price of the order",
                        "type": "number",
                        "format": "float",
                        "example": 150
                    },
                    "order_number": {
                        "description": "Unique order number",
                        "type": "string",
                        "example": "ORDER-123456"
                    },
                    "created_at": {
                        "description": "Order created at",
                        "type": "string",
                        "format": "date-time",
                        "example": "2021-08-01T00:00:00Z"
                    },
                    "updated_at": {
                        "description": "Order updated at",
                        "type": "string",
                        "format": "date-time",
                        "example": "2021-08-01T00:00:00Z"
                    },
                    "items": {
                        "description": "List of order items",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/OrderItem"
                        }
                    }
                },
                "type": "object"
            },
            "OrderItem": {
                "title": "OrderItem",
                "description": "OrderItem model",
                "properties": {
                    "id": {
                        "description": "ID of the order item",
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "order_id": {
                        "description": "ID of the order",
                        "type": "integer",
                        "example": 1
                    },
                    "book_id": {
                        "description": "ID of the book",
                        "type": "integer",
                        "example": 1
                    },
                    "quantity": {
                        "description": "Quantity of the book",
                        "type": "integer",
                        "example": 2
                    },
                    "price": {
                        "description": "Price of the book",
                        "type": "number",
                        "format": "float",
                        "example": 50
                    },
                    "created_at": {
                        "description": "Order item created at",
                        "type": "string",
                        "format": "date-time",
                        "example": "2021-08-01T00:00:00Z"
                    },
                    "updated_at": {
                        "description": "Order item updated at",
                        "type": "string",
                        "format": "date-time",
                        "example": "2021-08-01T00:00:00Z"
                    }
                },
                "type": "object"
            },
...
        },
...
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
    "security": [
        {
            "sanctum": []
        }
    ]
}